fixing test setup
