+++ js/plugins/firebase/package.json
@@ -22,7 +22,8 @@
     "build": "npm-run-all build:clean check compile",
     "build:watch": "tsup-node --watch",
     "jest": "node --experimental-vm-modules --dns-result-order=ipv4first node_modules/jest/bin/jest.js --runInBand --verbose",
-    "test": "(cd tests/functions && pnpm run build && cd .. && (firebase emulators:exec --only functions -P demo-test \"npm run jest\" 2>&1 | tee /tmp/firebase_integration_test))"
+    "test:build": "tsc ./tests/functions/index.ts --esModuleInterop --target es2017 --skipLibCheck --module commonjs --outDir .",
+    "test": "(pnpm run test:build && cd tests && (firebase emulators:exec --only functions -P demo-test \"npm run jest\" 2>&1 | tee /tmp/firebase_integration_test))"
   },
   "repository": {
     "type": "git",
@@ -53,7 +54,9 @@
     "@types/jest": "^29.5.12",
     "@jest/globals": "^29.7.0",
     "jest": "^29.7.0",
-    "ts-jest": "^29.1.2"
+    "ts-jest": "^29.1.2",
+    "firebase-functions-test": "^3.1.0",
+    "@genkit-ai/firebase": "workspace:*"
   },
   "types": "./lib/index.d.ts",
   "exports": {
