--- js/plugins/firebase/tests/functions/.env.local
@@ -1,1 +1,0 @@
-GENKIT_ENV=dev
