--- js/plugins/firebase/tests/functions/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "18"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "genkit": "workspace:*",
-    "@genkit-ai/firebase": "workspace:*",
-    "firebase-admin": "^12.6.0",
-    "firebase-functions": "^6.0.1"
-  },
-  "devDependencies": {
-    "@types/json-schema": "^7.0.15",
-    "firebase-functions-test": "^3.1.0",
-    "typescript": "^4.9.5",
-    "genkit-cli": "*",
-    "tsx": "^4.19.2"
-  },
-  "private": true
-}
