--- js/plugins/firebase/tests/functions/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2017",
-    "esModuleInterop": true,
-    "skipLibCheck": true
-  },
-  "compileOnSave": true,
-  "include": ["src"]
-}
